Extract cross cutting logic into `ModifierNode` and `ConditionalNode`
