+++ lib/rubocop.rb
@@ -21,6 +21,8 @@ require 'rubocop/node_pattern'
 require 'rubocop/string_interpreter'
 require 'rubocop/ast_node/sexp'
 require 'rubocop/ast_node'
+require 'rubocop/ast_node/node_extensions/conditional_node'
+require 'rubocop/ast_node/node_extensions/modifier_node'
 require 'rubocop/ast_node/node_extensions/array_node'
 require 'rubocop/ast_node/node_extensions/hash_node'
 require 'rubocop/ast_node/node_extensions/if_node'
