+++ lib/rubocop/ast_node/node_extensions/conditional_node.rb
@@ -1,0 +1,25 @@
+# frozen_string_literal: true
+
+module RuboCop
+  module NodeExtension
+    # Common functionality for nodes that have conditions:
+    # `if`, `while`, `until`, `case`
+    module ConditionalNode
+      def single_line_condition?
+        loc.keyword.line == condition.source_range.line
+      end
+
+      def multiline_condition?
+        !single_line_condition?
+      end
+
+      def condition
+        node_parts[0]
+      end
+
+      def body
+        node_parts[1]
+      end
+    end
+  end
+end
