+++ lib/rubocop/ast_node/node_extensions/until_node.rb
@@ -4,28 +4,15 @@ module RuboCop
   module NodeExtension
     # A node extension for `until` nodes.
     class UntilNode < RuboCop::Node
+      include ConditionalNode
+      include ModifierNode
+
       def keyword
         'until'
       end
 
       def do?
         loc.begin && loc.begin.is?('do')
-      end
-
-      def single_line_condition?
-        loc.keyword.line == condition.source_range.line
-      end
-
-      def multiline_condition?
-        !single_line_condition?
-      end
-
-      def condition
-        node_parts[0]
-      end
-
-      def body
-        node_parts[1]
       end
 
       def node_parts
