+++ lib/rubocop/ast_node/node_extensions/while_node.rb
@@ -4,28 +4,15 @@ module RuboCop
   module NodeExtension
     # A node extension for `while` nodes.
     class WhileNode < RuboCop::Node
+      include ConditionalNode
+      include ModifierNode
+
       def keyword
         'while'
       end
 
       def do?
         loc.begin && loc.begin.is?('do')
-      end
-
-      def single_line_condition?
-        loc.keyword.line == condition.source_range.line
-      end
-
-      def multiline_condition?
-        !single_line_condition?
-      end
-
-      def condition
-        node_parts[0]
-      end
-
-      def body
-        node_parts[1]
       end
 
       def node_parts
